move minimapEvents into minimap feature folder
